Allow allocating ranges with the size not being an exponential of 2
Since only a number of bits could be provided, only ranges where
the size was an exponential of 2 could be allocated. Since the
kernel already supports retyping untyped memory to ranges which are
a multiple of the minimal range size, the userspace now also
supports this by passing the correct parameters to the kernel.