Remove layer of indirection for context store
Instead of having different ranges for context and stack (which
were adjacent and had same access permissions in virtual memory
anyway), we store them in a single range. Due to the aforementioned
previous memory layout, this does not affect security/data access
but results in less address translation overhead.
Also, the contexts/stack are stored in an array instead of having
a single read-only list that points to the corresponding contexts.
This removes a layer of indirection.