Transform count macro into excl
Due to a change in semantics, we do not have an SCCount instruction
that counts the SecDivs having access to a specified address with
certain permissions anymore but we only check whether access is
exclusive (i.e., at most the supervisor and the currently executing
SecDiv have access with the specified permissions).
This change in semantics is reflected in renaming count to excl.